Reset timeout on receiving progress
Fundamentally there are two purposes to timeouts: avoiding DOS, and avoiding runaway code/bugs. After this commit, two timeouts exist, one for each. The DOS timeout is a universal timeout that is never reset. On the other hand, the runaway timeout is reset any time progress is received, because that shows that the worker has not run away.